Update pong.py
Originally we planned to make one raspberry pi called host, to be responsible for all the movement of the ball and the paddle. And the other raspberry pi is only for receiving the ball position and displaying it.

But after making the first version of the game and test running it, we have seen a large delay between the actual position of the ball and the displayed position on the non-host raspberry pi.